refactor(blorktools): index.html
