+++ packages/blorktools/src/asset_debugger/index.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Asset Debugger - Blorktools</title>
+    <script>
+        // Redirect to the scene directory
+        window.location.href = './scene/asset_debugger.html';
+    </script>
+    <style>
+        body {
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            background-color: #121212;
+            color: #e1e1e1;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            height: 100vh;
+            margin: 0;
+        }
+        .loading {
+            text-align: center;
+        }
+        h1 {
+            color: #5d8fff;
+        }
+    </style>
+</head>
+<body>
+    <div class="loading">
+        <h1>Asset Debugger</h1>
+        <p>Redirecting to the Asset Debugger tool...</p>
+        <p>If you are not redirected automatically, <a href="./scene/asset_debugger.html" style="color: #5d8fff;">click here</a>.</p>
+    </div>
+</body>
+</html>
