+++ packages/blorktools/src/asset_debugger/index.js
@@ -5,7 +5,7 @@
  * for use in other parts of the application or external imports.
  */
 
-import { init } from "./asset_debugger";
+import { init } from "./scene/asset_debugger";
 
 // Prevent default drag-and-drop behavior for the entire document
 function preventDefaultDragBehavior() {
@@ -38,3 +38,6 @@ function initializeDebugger() {
 
 // Initialize on script load
 initializeDebugger();
+
+// Export the init function
+export { init };
