+++ packages/blorktools/src/asset_debugger/scene/asset_debugger.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Asset Debugger - Blorktools</title>
-    <link rel="stylesheet" href="./styles.css">
-    <link rel="stylesheet" href="./loading-splash/loading-splash.css">
+    <link rel="stylesheet" href="../styles.css">
+    <link rel="stylesheet" href="../loading-splash/loading-splash.css">
     <script>
         // Get pinned state from localStorage before DOM is constructed
         function getPinnedState() {
