+++ packages/blorktools/src/index.html
@@ -74,7 +74,7 @@
     <div class="tool-card">
       <h3>Asset Debugger</h3>
       <p>Test and debug PBR texture sets with a simple drag-and-drop interface. Supports Base Color, ORM, and Normal map atlases.</p>
-      <a href="./asset_debugger/asset_debugger.html">Open Asset Debugger</a>
+      <a href="./asset_debugger/index.html">Open Asset Debugger</a>
     </div>
     
     <div class="tool-card">
